requirements.txt
playwright==1.49.0
google-genai>=0.1.0
python-dotenv==1.0.1
rich==13.9.4